GLA-761 Avoid DBUS panic when removing last 'name' being tracked
The DBUS code is used when an OVAL systemdunitproperty object attempts to collect info on a unit AND we are not running in the remote_helper.  In addition, when trying to collect info on a non-existing unit, an additional attempt will be made to get information.  This can lead to a case where we get notified to remove the last ‘name’ that we are tracking.  The existing code gets an an array out of bounds panic when this happens.

This change simply empties the array when removing the last item.  If not last item is uses the existing 'move later names up 1' behavior.